Added version function to setup.py
This function will return the git version of source, still considering
alternatives for windows or non git/ sed machines.

We can either use import git to get around sed requirement but is
difficult to get into the desired format we want, and requires another
dependancy prior to being able to run setup which could cause headaches
especially if they are windows users, they are probably not used to this
kind of thing. Other possibility is subprocess but we cant guarantee
that these commands will succeed either.