Added experimental donation function
This function is here to allow a database to donate its documents
including ID to another database, maybe for syncing or otherwise. We
return a list of documents synced so that they might be deleted or
verified by some other function.

I do not know what errors will be raised as of yet, so if something
fails the whole sync will stop from that point on insead of skipping for
now. However this will be required functionality soon, but only for dev
atm.