Added delete, and delete test, and donate delete
Donate will now delete donated entries, so this can be usefull for
database backups where the initial db is size restricted for some
reason. Towards this end deleteId has been implemented to delete an
entry by a given id. To check this functionality a unit-test has been
made to vet it, / to allow vetting.